+++ core/src/com/mygdx/wartowers/utils/Constants.java
@@ -2,8 +2,8 @@ package com.mygdx.wartowers.utils;
 
 
 public class Constants {
-    public static int APP_WIDTH = 480;
-    public static int APP_HEIGHT = 800;
+    public static float APP_WIDTH = 480;
+    public static float APP_HEIGHT = 800;
     public static final String BATTLEGROUND_JSON_PATH = "BattlegroundPresets/preset1.json";
     public static final String SKIN_COSMIC_PATH = "font_skins/comic/comic-ui.json";
